backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
